requirements.txt
Django==3.2.9
djangorestframework==3.14.0
requests==2.25.1
